+++ backend/__init__.py
@@ -1,0 +1,4 @@
+from .backend import Backend, get_backend
+from .picamera_backend import PiCameraBackend, ON_PI
+from .opencv_backend import OpenCVBackend
+__all__ = ["Backend", "get_backend", "PiCameraBackend", "ON_PI", "OpenCVBackend"]
